// generates/gql/graphql.ts
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
};
// ...
export type IBasicOptions = {
  __typename?: 'IBasicOptions';
  /** Контакты */
  contacts?: Maybe<IContacts>;
  /** Логотип */
  logo?: Maybe<ILogo>;
};

/** Контакты */
export type IContacts = {
  __typename?: 'IContacts';
  /** Адресс */
  address?: Maybe<Scalars['String']['output']>;
  /** Чистый телефон */
  clear_tel?: Maybe<Scalars['String']['output']>;
  /** Модифицированный телефон */
  modify_tel?: Maybe<Scalars['String']['output']>;
};

/** Тип иконки */
export type IImage = {
  __typename?: 'IImage';
  /** Алт иконки */
  alt?: Maybe<Scalars['String']['output']>;
  /** Путь до иконки */
  src?: Maybe<Scalars['String']['output']>;
  /** Заголовок иконки */
  title?: Maybe<Scalars['String']['output']>;
};

/** Логотип */
export type ILogo = {
  __typename?: 'ILogo';
  /** Путь до логотипа */
  src?: Maybe<Scalars['String']['output']>;
  /** Заголовок логотипа */
  title?: Maybe<Scalars['String']['output']>;
};

/** Метаданные о странице */
export type IMeta = {
  __typename?: 'IMeta';
  description?: Maybe<Scalars['String']['output']>;
  keywords?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};

/** Главная информация о странице */
export type IOptionsPage = {
  __typename?: 'IOptionsPage';
  backshow_slides?: Maybe<Array<Maybe<IImage>>>;
  /** Слоган страницы */
  slogan?: Maybe<Scalars['String']['output']>;
};
// ...
export type GetBasicOptionsQueryVariables = Exact<{ [key: string]: never; }>;


export type GetBasicOptionsQuery = { __typename?: 'RootQuery', basicOptions?: { __typename?: 'IBasicOptions', logo?: { __typename?: 'ILogo', src?: string | null, title?: string | null } | null, contacts?: { __typename?: 'IContacts', address?: string | null, clear_tel?: string | null, modify_tel?: string | null } | null } | null };

export type GetMetaPageQueryVariables = Exact<{
  slug?: InputMaybe<Scalars['String']['input']>;
}>;


export type GetMetaPageQuery = { __typename?: 'RootQuery', metaPage?: { __typename?: 'IMeta', title?: string | null, description?: string | null, keywords?: string | null } | null };

export type GetPageOptionsQueryVariables = Exact<{
  slug?: InputMaybe<Scalars['String']['input']>;
}>;


export type GetPageOptionsQuery = { __typename?: 'RootQuery', optionsPage?: { __typename?: 'IOptionsPage', slogan?: string | null, backshow_slides?: Array<{ __typename?: 'IImage', alt?: string | null, src?: string | null, title?: string | null } | null> | null } | null };
// ...
export const GetBasicOptionsDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"GetBasicOptions"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"basicOptions"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"logo"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"src"}},{"kind":"Field","name":{"kind":"Name","value":"title"}}]}},{"kind":"Field","name":{"kind":"Name","value":"contacts"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"address"}},{"kind":"Field","name":{"kind":"Name","value":"clear_tel"}},{"kind":"Field","name":{"kind":"Name","value":"modify_tel"}}]}}]}}]}}]} as unknown as DocumentNode<GetBasicOptionsQuery, GetBasicOptionsQueryVariables>;
// ...
export const GetPageOptionsDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"GetPageOptions"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"slug"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"optionsPage"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"slug"},"value":{"kind":"Variable","name":{"kind":"Name","value":"slug"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"slogan"}},{"kind":"Field","name":{"kind":"Name","value":"backshow_slides"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"alt"}},{"kind":"Field","name":{"kind":"Name","value":"src"}},{"kind":"Field","name":{"kind":"Name","value":"title"}}]}}]}}]}}]} as unknown as DocumentNode<GetPageOptionsQuery, GetPageOptionsQueryVariables>;